+++ src/main/java/com/colaborativesaving/demo/loans/controllers/contracts/RequestLoanType.java
@@ -5,9 +5,9 @@ import com.colaborativesaving.demo.loans.model.LoanType;
 public class RequestLoanType {
 
     private String name;
-    private int number;
-    private long installment;
-    private double interest;
+    private short number;
+    private double installment;
+    private float interest;
 
     public LoanType getType() {
         LoanType loanType = new LoanType();
@@ -24,9 +24,9 @@ public class RequestLoanType {
 
     public int getNumber() { return number; }
 
-    public void setNumber(int number) { this.number = number; }
+    public void setNumber(short number) { this.number = number; }
 
-    public long getInstallment() { return installment; }
+    public double getInstallment() { return installment; }
 
     public void setInstallment(long installment) { this.installment = installment; }
 
@@ -34,7 +34,5 @@ public class RequestLoanType {
         return interest;
     }
 
-    public void setInterest(double interest) {
-        this.interest = interest;
-    }
+    public void setInterest(float interest) { this.interest = interest; }
 }
